requirements.txt
streamlit==1.30.0
qrcode==7.4.2
Pillow==10.0.1
